# book-network-ui/src/app/services/models/borrowed-book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BorrowedBookResponse {
  author?: string;
  id?: number;
  isbn?: string;
  rating?: number;
  returnApproved?: boolean;
  returned?: boolean;
  title?: string;
}