# book-network/src/main/java/com/vnq/booknetwork/feedback/FeedbackResponse.java
package com.vnq.booknetwork.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {
    private Double note;
    private String comment;
    private boolean ownFeedback;

}